requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
python-dotenv==1.0.0


pytest==7.4.0
pytest-flask==1.2.0
coverage==7.2.7